+++ apps/api-server/src/middleware/security-headers.js
@@ -21,7 +21,7 @@ module.exports = function( req, res, next ) {
     res.header('Access-Control-Allow-Origin', url );
   }
   res.header('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE,OPTIONS,PATCH');
-  res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With, x-http-method-override, X-GRIP-Tenant-Id, X-Authorization');
+  res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With, x-http-method-override, X-GRIP-Tenant-Id');
   res.header('Access-Control-Allow-Credentials', 'true');
 
 	if (process.env.NODE_ENV != 'development') {