+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## UNRELEASED
+## All
 * Move API, auth and image servers, and the poc components, to one repo
+* Refactor: use normal Authorization header instead of X-Authorization to connect to the API
 
 ### API server
 * Add auth adapters and auth through third party IDPs - see /doc/auth.md